Decouple FF prompt path from cache/HTTP in PromptManager
When ff_prompt_serving is enabled, get_prompt() now tries FF evaluation
first without touching the cache layer. If FF succeeds, return
immediately (sub-ms, always fresh from the Agent). If FF returns nothing
(e.g. prompts created before the FFE write path), fall back to the HTTP
path with caching.

This separates the two resolution strategies cleanly: FF results are
never cached (the Agent's UFC cache is the source of truth), while HTTP
results continue to use hot/warm caching as before.

Also: consolidate FF tests into test_prompts.py, remove
test_ff_prompt_enable.py and test_ff_prompt_evaluation.py.
